+++ src/main/java/org/folio/service/transformationfields/TransformationFieldsConfig.java
@@ -1,14 +1,23 @@
-package org.folio.service.fieldname;
+package org.folio.service.transformationfields;
 
 import java.util.Map;
 
 import static org.folio.service.mapping.referencedata.ReferenceDataImpl.IDENTIFIER_TYPES;
 
+/**
+ * Initial data for the transformation field. While extending the enum, put new values in alphabetical order
+ */
 public enum TransformationFieldsConfig {
 
+  ID("id", "$.{recordType}.id"),
+  HR_ID("hrid", "$.{recordType}.hrid"),
+  INSTANCE_ID("instanceId", "$.holdings.instanceId"),
   IDENTIFIERS("identifiers", "$.{recordType}[*].identifiers[?(@identifierTypeId=={id})].value", IDENTIFIER_TYPES),
-  FIXED_LENGTH_DATA_ELEMENT("metadata.createdDate", "$.{recordType}.metadata.createdDate", MetadataParametersConstants.getFixedLengthDataElement()),
-  HR_ID("hrid", "$.{recordType}.hrid");
+
+  METADATA_CREATED_DATE("metadata.createdDate", "$.{recordType}.metadata.createdDate", MetadataParametersConstants.getFixedLengthDataElement()),
+  METADATA_UPDATED_DATE("metadata.updatedDate", "$.{recordType}.metadata.updatedDate"),
+  METADATA_CREATED_BY_USER_ID("metadata.createdByUserId", "$.{recordType}.metadata.createdByUserId"),
+  METADATA_UPDATED_BY_USER_ID("metadata.updatedByUserId", "$.{recordType}.metadata.updatedByUserId");
 
   private final String fieldId;
   private final String path;