+++ src/main/java/org/folio/service/transformationfields/builder/JsonPathBuilder.java
@@ -1,7 +1,7 @@
-package org.folio.service.fieldname.builder;
+package org.folio.service.transformationfields.builder;
 
 import org.folio.rest.jaxrs.model.TransformationField.RecordType;
-import org.folio.service.fieldname.TransformationFieldsConfig;
+import org.folio.service.transformationfields.TransformationFieldsConfig;
 import org.springframework.stereotype.Component;
 
 @Component