+++ NEWS.md
@@ -1,3 +1,11 @@
+## 07/29/2020 v2.1.4 Released
+This is a bugfix release to address changes missing metadata for default job profiles and mapping profiles
+
+[Full Changelog](https://github.com/folio-org/mod-data-export/compare/v2.1.3...v2.1.4)
+
+### Bug Fixes
+* [MDEXP-253](https://issues.folio.org/browse/MDEXP-253) - Default Profiles do not have metadata causing error screen
+
 ## 07/28/2020 v2.1.3 Released
 This is a bugfix release to address upgrade issues from v1.1.1 to current version because of the migration scripts
 
@@ -5,6 +13,7 @@ This is a bugfix release to address upgrade issues from v1.1.1 to current versio
 
 ### Bug Fixes
 * [MDEXP-247](https://issues.folio.org/browse/MDEXP-247) - Data Migration issue on goldenrod
+
 
 ## 07/13/2020 v2.1.2 Released
 This bugfix release includes fixes to concurrency issues while exporting jobs and also fixes empty query sent to